# Conflicts:
#	README.md
#	docs/IE_DG/API_Changes.md
#	docs/IE_DG/Samples_Overview.md
#	docs/MO_DG/prepare_model/Supported_Frameworks_Layers.md
#	docs/MO_DG/prepare_model/convert_model/Convert_Model_From_TensorFlow.md
#	docs/MO_DG/prepare_model/convert_model/pytorch_specific/Convert_F3Net.md
#	docs/ops/detection/ExperimentalDetectronDetectionOutput_6.md
#	docs/ops/detection/ExperimentalDetectronGenerateProposalsSingleImage_6.md
#	docs/ops/detection/ExperimentalDetectronPriorGridGenerator_6.md
#	docs/ops/detection/ExperimentalDetectronROIFeatureExtractor_6.md
#	docs/ops/sort/ExperimentalDetectronTopKROIs_6.md
#	inference-engine/cmake/vpu_dependencies.cmake
#	inference-engine/ie_bridges/c/samples/hello_classification/README.md
#	inference-engine/ie_bridges/c/samples/hello_nv12_input_classification/README.md
#	inference-engine/ie_bridges/c/samples/object_detection_sample_ssd/README.md
#	inference-engine/ie_bridges/python/sample/classification_sample_async/README.md
#	inference-engine/ie_bridges/python/sample/hello_classification/README.md
#	inference-engine/ie_bridges/python/sample/hello_reshape_ssd/README.md
#	inference-engine/ie_bridges/python/sample/object_detection_sample_ssd/README.md
#	inference-engine/ie_bridges/python/wheel/meta/openvino-dev.setup.cfg
#	inference-engine/samples/classification_sample_async/README.md
#	inference-engine/samples/hello_classification/README.md
#	inference-engine/samples/hello_nv12_input_classification/README.md
#	inference-engine/samples/hello_query_device/README.md
#	inference-engine/samples/hello_reshape_ssd/README.md
#	inference-engine/samples/object_detection_sample_ssd/README.md
#	inference-engine/samples/speech_sample/README.md
#	inference-engine/samples/style_transfer_sample/README.md
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_bin_conv_node.cpp
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_def_conv_node.cpp
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_pad_node.cpp
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_pad_node.h
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_quantize_node.cpp
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_roi_pooling_node.cpp
#	inference-engine/src/mkldnn_plugin/nodes/mkldnn_split_node.cpp
#	model-optimizer/extensions/front/tf/assign_elimination.py
#	model-optimizer/mo/middle/passes/fusing/helpers.py
#	model-optimizer/mo/middle/passes/fusing/helpers_test.py
#	model-optimizer/telemetry/utils/sender_test.py
#	ngraph/core/reference/src/runtime/reference/loop.cpp
#	ngraph/python/src/pyngraph/node_factory.cpp
#	ngraph/python/tests/test_ngraph/test_create_op.py
#	ngraph/test/util.cpp

+++ ngraph/core/reference/include/ngraph/runtime/reference/gather_elements.hpp
@@ -1,18 +1,6 @@
-//*****************************************************************************
-// Copyright 2017-2021 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//*****************************************************************************
 
 #pragma once
 
@@ -38,7 +26,7 @@ namespace ngraph
                 {
                     axis += data_shape.size();
                 }
-                if (axis < 0 || axis >= data_shape.size())
+                if (axis < 0 || axis >= static_cast<int64_t>(data_shape.size()))
                 {
                     throw std::domain_error{
                         "axis for GatherElements exceeds allowed range [0, data_rank)"};
@@ -47,9 +35,9 @@ namespace ngraph
                 // in 1D case results can be achieved without additional calculations
                 if (data_shape.size() == 1)
                 {
-                    for (int64_t i = 0; i < indices_shape[0]; i++)
+                    for (size_t i = 0; i < indices_shape[0]; i++)
                     {
-                        if (indices[i] >= data_shape[0])
+                        if (static_cast<size_t>(indices[i]) >= data_shape[0])
                         {
                             throw std::domain_error{
                                 "indices values of GatherElement exceed data size"};
@@ -66,11 +54,11 @@ namespace ngraph
                 size_t data_num_columns = data_shape[1];
                 if (data_shape.size() == 2)
                 {
-                    int64_t idx;
+                    size_t idx;
                     if (axis == 0)
                     {
-                        for (int64_t i = 0; i < num_rows; i++)
-                            for (int64_t j = 0; j < num_columns; j++)
+                        for (size_t i = 0; i < num_rows; i++)
+                            for (size_t j = 0; j < num_columns; j++)
                             {
                                 idx = indices[num_columns * i + j];
                                 if (idx < 0 || idx >= data_shape[0])
@@ -84,8 +72,8 @@ namespace ngraph
                     }
                     else // axis == 1
                     {
-                        for (int64_t i = 0; i < num_rows; i++)
-                            for (int64_t j = 0; j < num_columns; j++)
+                        for (size_t i = 0; i < num_rows; i++)
+                            for (size_t j = 0; j < num_columns; j++)
                             {
                                 idx = indices[num_columns * i + j];
                                 if (idx < 0 || idx >= data_shape[1])
@@ -120,13 +108,13 @@ namespace ngraph
                 */
 
                 size_t max_inner_sum = 1;
-                for (int i = axis + 1; i < indices_shape.size(); i++)
+                for (size_t i = axis + 1; i < indices_shape.size(); i++)
                     max_inner_sum *= indices_shape[i];
 
                 size_t max_outer_sum = 1, outer_sum_inc = 1;
                 for (int i = 0; i < axis; i++)
                     max_outer_sum *= indices_shape[i];
-                for (int i = axis; i < data_shape.size(); i++)
+                for (size_t i = axis; i < data_shape.size(); i++)
                     outer_sum_inc *= data_shape[i];
                 max_outer_sum *= outer_sum_inc;
 
@@ -135,7 +123,8 @@ namespace ngraph
                     for (size_t k = 0; k < indices_shape[axis]; k++)
                         for (size_t inner_sum = 0; inner_sum < max_inner_sum; inner_sum++)
                         {
-                            if (indices[i] < 0 || indices[i] >= data_shape[axis])
+                            if (indices[i] < 0 ||
+                                static_cast<size_t>(indices[i]) >= data_shape[axis])
                             {
                                 throw std::domain_error{
                                     "indices values of GatherElement exceed data size"};
