+++ ngraph/frontend/onnx_import/src/op/dequantize_linear.cpp
@@ -1,18 +1,6 @@
-//*****************************************************************************
-// Copyright 2017-2021 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//*****************************************************************************
 
 #include <cstdint>
 #include <memory>
@@ -142,7 +130,7 @@ namespace ngraph
                     {
                         std::vector<int64_t> target_dims;
 
-                        for (size_t i = 0; i < axis; ++i)
+                        for (int64_t i = 0; i < axis; ++i)
                         {
                             target_dims.push_back(1);
                         }
@@ -157,7 +145,7 @@ namespace ngraph
                             target_dims.push_back(0);
                         }
 
-                        for (size_t i = axis + 1; i < x_shape.rank().get_length(); ++i)
+                        for (int64_t i = axis + 1; i < x_shape.rank().get_length(); ++i)
                         {
                             target_dims.push_back(1);
                         }
