+++ ngraph/test/type_prop/split.cpp
@@ -1,18 +1,6 @@
-//*****************************************************************************
-// Copyright 2017-2021 Intel Corporation
+// Copyright (C) 2018-2021 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//*****************************************************************************
 
 #include "gtest/gtest.h"
 #include "ngraph/ngraph.hpp"
@@ -90,7 +78,7 @@ TEST(type_prop, split_v1)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_element_type(i), element::f16);
         EXPECT_EQ(split->get_output_shape(i), (Shape{2, 1, 4}));
@@ -106,7 +94,7 @@ TEST(type_prop, split_v1_axis_const_data_axis_dim_known)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), (PartialShape{2, 1, Dimension::dynamic()}));
     }
@@ -121,7 +109,7 @@ TEST(type_prop, split_v1_axis_const_only_data_axis_dim_known)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i),
                   (PartialShape{1, Dimension::dynamic(), Dimension::dynamic()}));
@@ -137,7 +125,7 @@ TEST(type_prop, split_v1_axis_const_data_axis_dim_unknown)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i),
                   (PartialShape{4, Dimension::dynamic(), 3, 5}));
@@ -152,7 +140,7 @@ TEST(type_prop, split_v1_axis_const_only_data_rank_known)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), PartialShape::dynamic(4));
     }
@@ -166,7 +154,7 @@ TEST(type_prop, split_v1_axis_not_const_only_data_rank_known)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), PartialShape::dynamic(4));
     }
@@ -180,7 +168,7 @@ TEST(type_prop, split_v1_axis_const_data_rank_unknown)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), PartialShape::dynamic());
     }
@@ -194,7 +182,7 @@ TEST(type_prop, split_v1_axis_not_const_data_rank_unknown)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), PartialShape::dynamic());
     }
@@ -208,7 +196,7 @@ TEST(type_prop, split_v1_axis_dynamic_rank)
     const auto split = make_shared<op::v1::Split>(data, axis, num_splits);
 
     EXPECT_EQ(split->outputs().size(), num_splits);
-    for (int i = 0; i < num_splits; ++i)
+    for (size_t i = 0; i < num_splits; ++i)
     {
         EXPECT_EQ(split->get_output_partial_shape(i), PartialShape::dynamic());
     }
